% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mk_date_parse.R
\name{mk_date_parse}
\alias{mk_date_parse}
\title{This is a helper function to detect the date format of a character vector with dates}
\usage{
mk_date_parse(date)
}
\arguments{
\item{character}{vector of dates}
}
\description{
This is a helper function to detect the date format of a character vector with dates
}

