% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_date_parse.R
\name{ws_date_parse}
\alias{ws_date_parse}
\title{This is a helper function to detect the date format of a character vector with dates}
\usage{
ws_date_parse(date)
}
\arguments{
\item{character}{vector of dates}
}
\description{
This is a helper function to detect the date format of a character vector with dates
}
