% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_download_file.R
\name{ws_download_file}
\alias{ws_download_file}
\title{Function to download a file from MKonline}
\usage{
ws_download_file(key, area, ws_user = Sys.getenv("WS_USER"),
  ws_password = Sys.getenv("WS_PASSWORD"))
}
\arguments{
\item{key}{The key (as in file) to download (ex CON_POW_H_A)}

\item{area}{The area (ex: cee, cwe, np, it etc.)}

\item{ws_user}{Your Wattsight user name. Defaults to an Environment file called WS_USER.}

\item{ws_password}{Your Wattsight user password. Defaults to an Environment file called WS_PASSWORD.}
}
\description{
Function to download a file from MKonline
}
