% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_list_files.R
\name{ws_list_files}
\alias{ws_list_files}
\title{Function to return the file list.}
\usage{
ws_list_files(only_accessible = FALSE, ws_user = Sys.getenv("WS_USER"),
  ws_password = Sys.getenv("WS_PASSWORD"))
}
\arguments{
\item{only_accessible}{TRUE if you only want to see your available data series.}

\item{ws_user}{Your Wattsight user name. Defaults to a environment variable called WS_USER.}

\item{ws_password}{Your Wattsight user password. Defaults to a environment variable called WS_USER.}
}
\description{
The files list can be returned with either all the files or only the accessible files.
}
