% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_download_file.R
\name{mk_download_file}
\alias{mk_download_file}
\title{Function to download a file from MKonline}
\usage{
mk_download_file(mk_user, mk_password, key, area)
}
\arguments{
\item{mk_user}{Your MKOnline user name.}

\item{mk_password}{Your MKOnline user password.}

\item{key}{The key (as in file) to download (ex CON_POW_H_A)}

\item{area}{The area (ex: cee, cwe, np, it etc.)}
}
\description{
Function to download a file from MKonline
}

