% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mk_download_file}
\alias{mk_download_file}
\title{Function to download a file from MKonline}
\usage{
mk_download_file(mk_user, mk_password, mk_file, mk_area)
}
\arguments{
\item{mk_user}{}

\item{mk_password}{}

\item{mk_file}{The key (as in file) to download (ex CON_POW_H_A)}

\item{mk_area}{The key from area (ex: cee, cwe, np, it etc.)}
}
\description{
Function to download a file from MKonline
}

