% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_list_files.R
\name{mk_list_files}
\alias{mk_list_files}
\title{Function to return the file list}
\usage{
mk_list_files(mk_user, mk_password, only_accessible = FALSE,
  time_zone = "GMT")
}
\arguments{
\item{mk_user}{Your MKOnline user name.}

\item{mk_password}{Your MKOnline user password.}

\item{only_accessible}{TRUE if you only want to see your available data series.}

\item{time_zone}{GMT (UTC) or another like CET. Used to parse the updated column}
}
\description{
The files list can be returned with either all the files or only the accessible files.
}

